utils: Implement a tool to lock/unlock u-boot console
This is a userspace program to lock and unlock the u-boot console,
by modifying the u-boot env area in the MBR of the SD image.

If no argument given, it shows the current lock state. Pass
the argument -l to lock or -u to unlock.

To setup a chameleon board, after the MAC address is set, better
to execute this program "lock_u_boot_console -l" to lock the console.

If we need to change the MAC address again, first execute this
program "lock_u_boot_console -u" to unlock the console.

BUG=b:70692765
TEST=Run ""lock_u_boot_console", "lock_u_boot_console -l", and
"lock_u_boot_console -u" on Chameleon.

Change-Id: Ide376805a3b8c0460edf978b0bbf8db437770fa4
Reviewed-on: https://chromium-review.googlesource.com/875179
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>