Add UnplugHPD method to video ports
i915 is using chamelium boards to simulate some real life scenarios
and for this scenario we need to unplug only the HDP leaving EDID and
DDC connected but the Unplug() method disables all three.

So adding this new method that will only disconect the HPD and
as it is a new method it will not break any test that uses Plug().

Change-Id: Ic9c6d872527b17070db0193f19a202e3202c6b87
Reviewed-on: https://chromium-review.googlesource.com/1530030
Commit-Ready: Jose Souza <jose.souza@intel.com>
Tested-by: Jose Souza <jose.souza@intel.com>
Reviewed-by: Jose Souza <jose.souza@intel.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>