autotest: fix nRF52 error message on missing authentication at reboot
Per reset cycle of the nRF52 Noridc LE device an error messages was printed.
While fixing this, it turns out that this exception avoided a device reboot
that was part of the original Init flow in bluetooth_hid.py.
Reseting of nRF52 is not needed, this change also avoids it for that
platform.

BUG=chromium:949394
TEST=running the test on eve device

Change-Id: Ic94c71c20440ec93612e151525d4866e0ba241be
Reviewed-on: https://chromium-review.googlesource.com/1553964
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Commit-Ready: Yoni Shavit <yshavit@chromium.org>
Tested-by: Yoni Shavit <yshavit@chromium.org>
Reviewed-by: Neeraj Poojary <npoojary@google.com>